Close to having a working background tile
- Improved `Makefile` to support building multiple C files
- Added a rendering file

The current issue appears to be with the TileMapping tool I have initially
picked. The tool seems incredibly unstable and unusable (at least on Linux).

Building tilemaps appears to work initially but the files generated from
Tile Map Studio cannot be re-opened by the program. It just loads in an empty
map.

Loading multiple tiles at once is a PITA. The file browser doesn't remember
where you left off and there appears to be no "bulk upload".

Running the program on a 4K monitor requires a lot of squinting. The UI does
not want to scale up when you try to zoom in.

Given all that, I believe the export function is dumping crap (based on the
broken image now displayed by the game) to the C file. And annoyingly, it
gives only a C file so I have to add my own header and header guards.

Overall, I've spent most of my Saturday fighting with the tile map tool and I
think I need to find a better one for Linux.